+++ hello-world/src/components/home/helloWorldSnippets/controller.tsx
@@ -1,20 +1,18 @@
 "use client";
 
-import { Dispatch, memo, MutableRefObject, ReactElement, Reducer, useCallback, useEffect, useReducer, useState } from "react";
+import { memo, MutableRefObject, ReactElement, Reducer, useCallback, useEffect, useReducer, useState } from "react";
 import { useGSAP } from "@gsap/react";
 import gsap from "gsap";
 import PropTypes from "prop-types";
 
 import { withDisplayName } from "@/ui/decorator";
-import { PaginationActions, PaginationDispatcherAction, PaginationPage, useToggle } from "@/ui/hooks";
+import { PaginationActions, useToggle } from "@/ui/hooks";
+import { CarouselControllerProps } from "@/ui/widgets/carousel";
 import { ICON_NAMES, IconButton } from "@/ui/widgets/icon";
 
 const STATUS = Object.freeze({
-    NEXT: "next",
-    NONE: "none",
     PAUSE: "pause",
     PLAY: "play",
-    PREVIOUS: "previous",
     RESTART: "restart",
     STOP: "stop"
 });
@@ -32,114 +30,105 @@ const INITIAL_CONTROLLER_STATE: State = {
     showPause: false,
     showPlay: true,
     showRestart: false,
-    status: STATUS.NONE
+    status: STATUS.STOP
 } as const;
 
-function useGAPAnimation(pageItemRef: MutableRefObject<HTMLDivElement[]>, toggleIsTweenActive): [gsap.core.Tween, () => void] {
-    const [reloadId, setReloadId] = useState(null);
+function useGAPAnimation(containerRef?: MutableRefObject<HTMLDivElement>, containerItemClassName?: string): [() => void, [gsap.core.Tween | null, boolean]] {
+    const [isTweenActive, toggleIsTweenActive] = useToggle(false);
     const [tween, setTween] = useState<gsap.core.Tween | null>(null);
-    useGSAP(() => {
-        const gsapTween = gsap.fromTo(pageItemRef.current!,
-            {
-                y: 100,
-                x: 0,
-                opacity: 0
-            },
-            {
-                y: 0,
-                opacity: 1,
-                delay: 1,
-                duration: 7,
-                onStart: () => {
-                    toggleIsTweenActive(true);
-                },
-                onComplete: () => {
-                    toggleIsTweenActive(false);
-                }
-            }
-        );
-        setTween(gsapTween);
-    }, [reloadId])
+    const [reloadId, setReloadId] = useState(null);
+
     const animate = useCallback(() => {
         setReloadId(Symbol("UI.HELLO-WORLD-SNIPPETS.USE-GAP-ANIMATION.RELOAD-ID"));
     }, []);
-    return [tween, animate];
+
+    useGSAP(() => {
+        if(reloadId) {
+            const createdTween = gsap.fromTo(`.${containerItemClassName}`,
+                {
+                    y: 100,
+                    x: 0,
+                    opacity: 0
+                },
+                {
+                    y: 0,
+                    opacity: 1,
+                    delay: 1,
+                    duration: 7,
+                    onStart: () => {
+                        toggleIsTweenActive(true);
+                    },
+                    onComplete: () => {
+                        toggleIsTweenActive(false);
+                    }
+                }
+            );
+            setTween(createdTween);
+        }
+    }, {dependencies: [reloadId], revertOnUpdate: true, scope: containerRef});
+
+    return [animate, [tween, isTweenActive]];
 }
 
 function controllerStateReducer(previousState: State, action: ControllerStatus): State {
-    if(action === STATUS.PREVIOUS) {
-        // clicking this should do the ff:
-        // 1. display the previously displayed content
-        // 2. display the pause button
-        return {...previousState, status: STATUS.PREVIOUS, showPause: true, showPlay: false, showRestart: false};
-    }
-    if(action === STATUS.NEXT) {
-        // clicking this should do the ff:
-        // 1. display the next content
-        // 2. display the pause button
-        return {...previousState, status: STATUS.NEXT, showPause: true, showPlay: false, showRestart: false};
-    }
-    if(action === STATUS.NONE) {
-        return INITIAL_CONTROLLER_STATE;
-    }
     if(action === STATUS.PAUSE) {
-        // clicking this should do the ff:
-        // 1. pause the display of subsequent contents
-        // 2. display the play button
-        return {...previousState, status: STATUS.PAUSE, showPause: false, showPlay: true, showRestart: false};
+        return {
+            ...previousState,
+            showPause: false,
+            showPlay: true,
+            showRestart: false,
+            status: STATUS.PAUSE
+        };
     }
     if(action === STATUS.PLAY) {
-        // clicking this should do the ff:
-        // 1. conitnue with displaying subseqent contents
-        // 2. display the pause button
-        return {...previousState, status: STATUS.PLAY, showPause: true, showPlay: false, showRestart: false};
+        return {
+            ...previousState,
+            showPause: true,
+            showPlay: false,
+            showRestart: false,
+            status: STATUS.PLAY
+        };
     }
     if(action === STATUS.RESTART) {
-        // clicking this should do the ff:
-        // 1. restart displaying contents from the first
-        // 2. display pause button
-        return {...previousState, status: STATUS.RESTART, showPause: true, showPlay: false, showRestart: false};
+        return {
+            ...previousState,
+            showPause: false,
+            showPlay: false,
+            showRestart: true,
+            status: STATUS.RESTART
+        };
     }
     if(action === STATUS.STOP) {
-        // clicking this should do the ff:
-        // 1. stop/kill a particular displayed content's tween
-        // 2. display the restart button
-        return {...previousState, status: STATUS.STOP, showPause: false, showPlay: false, showRestart: true};
+        return INITIAL_CONTROLLER_STATE;
     }
     return previousState;
 }
 
-type Props = {
-    onPageChange?: Dispatch<PaginationDispatcherAction>;
-    page?: PaginationPage;
-    pageItemRef?: MutableRefObject<HTMLDivElement[]>;
-};
-
 Controller.propTypes = {
+    containerProps: PropTypes.shape({
+        ref: PropTypes.shape({
+            current: PropTypes.node
+        }),
+        itemClassName: PropTypes.string
+    }),
     onPageChange: PropTypes.func,
     page: PropTypes.shape({
         count: PropTypes.number,
         current: PropTypes.number
-    }),
-    pageItemRef: PropTypes.shape({
-        current: PropTypes.array
     })
 };
 
 function Controller({
+    containerProps,
     onPageChange,
-    page,
-    pageItemRef
-}: Props): ReactElement<Props> {
+    page
+}: CarouselControllerProps): ReactElement<CarouselControllerProps> {
     const [controllerState, dispatchControllerStateChange] = useReducer<Reducer<State, ControllerStatus>>(controllerStateReducer, INITIAL_CONTROLLER_STATE);
-    const [isTweenActive, toggleIsTweenActive] = useToggle(false);
-    const [tween, animate] = useGAPAnimation(pageItemRef, toggleIsTweenActive);
+    const [animate, [tween, isTweenActive]] = useGAPAnimation(containerProps?.ref, containerProps?.itemClassName);
 
     const handleNextClick = useCallback(() => {
-        dispatchControllerStateChange(STATUS.NEXT);
         onPageChange?.({type: PaginationActions.NEXT});
-        animate();
-    }, [animate, onPageChange]);
+    }, [onPageChange]);
 
     const handlePauseClick = useCallback(() => {
         dispatchControllerStateChange(STATUS.PAUSE);
@@ -152,36 +141,35 @@ function Controller({
     }, [tween]);
 
     const handlePreviousClick = useCallback(() => {
-        dispatchControllerStateChange(STATUS.PREVIOUS);
         onPageChange?.({type: PaginationActions.PREVIOUS});
-        animate();
-    }, [animate, onPageChange]);
+    }, [onPageChange]);
 
     const handleRestartClick = useCallback(() => {
-        dispatchControllerStateChange(STATUS.RESTART);
         onPageChange?.({type: PaginationActions.RESET});
+    }, [onPageChange]);
+
+    useEffect(() => {
         animate();
-    }, [animate, onPageChange]);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+    }, [page]);
 
     useEffect(() => {
         if(isTweenActive) {
             dispatchControllerStateChange(STATUS.PLAY);
+        } else {
+            if(!page.isLast) {
+                handleNextClick();
+                dispatchControllerStateChange(STATUS.STOP);
+            } else {
+                dispatchControllerStateChange(STATUS.RESTART);
+            }
         }
-        if(!isTweenActive && page?.current !== page?.count) {
-            handleNextClick();
-        }
-        if(!isTweenActive && page?.current === page?.count) {
-            dispatchControllerStateChange(STATUS.STOP);
-        }
-        return () => {
-            dispatchControllerStateChange(STATUS.NONE);
-        };
     // eslint-disable-next-line react-hooks/exhaustive-deps
     }, [isTweenActive]);
 
     return (
         <div className="inline-flex justify-center">
-            <IconButton icon={ICON_NAMES.FAST_REWIND} onClick={handlePreviousClick} />
+            <IconButton disabled={!page || page.isFirst} icon={ICON_NAMES.FAST_REWIND} onClick={handlePreviousClick} />
             {
                 controllerState.showPause &&
                 <IconButton icon={ICON_NAMES.PAUSE_CIRCLE} onClick={handlePauseClick} />
@@ -194,7 +182,7 @@ function Controller({
                 controllerState.showRestart &&
                 <IconButton icon={ICON_NAMES.RESTART_ALT} onClick={handleRestartClick} />
             }
-            <IconButton icon={ICON_NAMES.FAST_FORWARD} onClick={handleNextClick} />
+            <IconButton disabled={!page || page.isLast} icon={ICON_NAMES.FAST_FORWARD} onClick={handleNextClick} />
         </div>
     );
 }