+++ hello-world/src/components/navigation/index.tsx
@@ -1,6 +1,6 @@
 export {default as NavigationList} from "./navigationList";
 export {default as NavigationFAB} from "./navigationFAB";
-export {NAVIGATION} from "./navigation.constants";
-export {default as useNavigation} from "./useNavigation.hook";
+export {NavigationLabel} from "./navigation.constants";
+export {default as useNavigation} from "./hooks/useNavigation";
 
-export type { NavigationType } from "./navigation.types";
+export type { NavigationLabelType } from "./navigation.types";