+++ hello-world/src/ui/context/configuration/configuration.constants.ts
@@ -1,9 +1,12 @@
+import {getDeviceType} from "@/ui/utils/device/";
+
 export const ThemeVariants = {
     DARK: "dark",
     LIGHT: "light"
 } as const;
 
 export const InitialConfiguration = {
+    deviceType: getDeviceType(),
     theme: {
         name: "Helloworld",
         variant: {