+++ hello-world/src/ui/utils/device/device.type.ts
@@ -1,0 +1,3 @@
+import {DeviceTypes as DeviceTypesConstant} from "./device.constants";
+
+export type DeviceType = typeof DeviceTypesConstant[keyof typeof DeviceTypesConstant];