+++ hello-world/src/ui/utils/device/index.ts
@@ -1,0 +1,4 @@
+export {default as getDeviceType} from "./getDeviceType";
+export {DeviceTypes} from "./device.constants";
+
+export type {DeviceType} from "./device.type";