+++ hello-world/src/ui/widgets/carousel/index.ts
@@ -1,1 +1,2 @@
 export {default as Carousel} from "./carousel";
+export type {Props as CarouselControllerProps} from "./carouselController";