+++ hello-world/src/ui/widgets/title/index.ts
@@ -1,2 +1,2 @@
 export { default as Title } from "./title";
-export type { Variant } from "./title";
+export type { Props as TitlePropsType, Variant } from "./title";